Add footer, hero section, and language switcher components
- Created Footer component with styling and links for About, Contact, and Privacy pages.
- Implemented HeroSection component featuring a video background and call-to-action buttons.
- Added LanguageSwitcher component to allow users to switch between English, Romanian, and Russian languages.
- Developed ThemeToggle component for light/dark mode switching.
- Introduced button component with various styles and sizes.
- Set up i18n for internationalization with translations for multiple languages.
- Created Navbar component with responsive design, including theme toggle and language switcher.
- Added pages for About, Contact, Home, Privacy, Products, and Login with corresponding styles.
- Implemented product listing with search and sorting functionality.
- Created registration and login forms with validation and error handling.
- Updated index.js to include i18n configuration.

+++ backend-django/shop/models.py
@@ -15,7 +15,7 @@ class Course(models.Model):
     title = models.CharField(max_length=255)
     price = models.FloatField()
     students_qty = models.IntegerField()
-    reviews_qty = models.IntegerField()
+    reviews_qty = models.FloatField()
     category = models.ForeignKey(Category, on_delete=models.CASCADE)
     created_at = models.DateTimeField(default=timezone.now)
